project: fix 'oauth2_provider' app scopes settings
